work on sh plots & bug fixes in sh optimizer
